+++ CHANGELOG.rst
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a
 Changelog <https://keepachangelog.com/en/1.1.0/>`__.
 
+Unreleased
+----------
+
+Added
+~~~~~
+
+- Add ``fields`` parameter to ``get_dataset_files``
+
 1.1.1 - 2025-08-20
 ------------------